fix - #101 splashStickerLabel 위치 수정
