+++ client/src/components/molecules/stickerLabel/StickerLabel.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Sticker from '../atoms/icons';
+import Sticker from '../../atoms/icons';
 
 const Container = styled.div`
   box-sizing: border-box;
